# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SDDFreak Requirements API",
    "description": "Internal API for SPEC-001 (Requirements Management System: Core). Internal per §2.1: documented and stable enough for the UI, with no third-party compatibility guarantee.",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/projects/{projectId}/requirements": {
      "get": {
        "summary": "List, filter, sort, and search requirements (FR-SRCH-001…006)",
        "tags": [
          "requirements"
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "string"
                }
              ]
            },
            "in": "query",
            "name": "type",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "string"
                }
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "string"
                }
              ]
            },
            "in": "query",
            "name": "priority",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "string"
                }
              ]
            },
            "in": "query",
            "name": "tag",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "categoryId",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "includeDescendants",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "uncategorized",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "ownerId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "authorId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "includeDeleted",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "key"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "title"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "status"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "priority"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "createdAt"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "updatedAt"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "asc"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "desc"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "direction",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "minimum": 0,
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "total",
                    "limit",
                    "offset"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Requirement"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "offset": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a requirement (FR-REQ-001, FR-REQ-002)",
        "tags": [
          "requirements"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "statement",
                  "type"
                ],
                "properties": {
                  "title": {
                    "minLength": 1,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "statement": {
                    "minLength": 1,
                    "maxLength": 10000,
                    "type": "string"
                  },
                  "type": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "functional"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "non_functional"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "technical"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "constraint"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "business"
                        ]
                      }
                    ]
                  },
                  "rationale": {
                    "anyOf": [
                      {
                        "maxLength": 10000,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "priority": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "must"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "should"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "could"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "wont"
                        ]
                      }
                    ]
                  },
                  "status": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "draft"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "proposed"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "approved"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "implemented"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "verified"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "rejected"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "obsolete"
                        ]
                      }
                    ]
                  },
                  "acceptanceCriteria": {
                    "anyOf": [
                      {
                        "maxLength": 10000,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "source": {
                    "anyOf": [
                      {
                        "maxLength": 200,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "ownerId": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "categoryId": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "parentId": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "tags": {
                    "maxItems": 50,
                    "type": "array",
                    "items": {
                      "maxLength": 50,
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requirement"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/requirements/tree": {
      "get": {
        "summary": "Category hierarchy with parent–child decomposition (FR-SRCH-007)",
        "tags": [
          "requirements"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/projects/{projectId}/requirements/export": {
      "get": {
        "summary": "Export the filtered result set as CSV or JSON (FR-SRCH-009)",
        "tags": [
          "requirements"
        ],
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "string"
                }
              ]
            },
            "in": "query",
            "name": "type",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "string"
                }
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "string"
                }
              ]
            },
            "in": "query",
            "name": "priority",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "string"
                }
              ]
            },
            "in": "query",
            "name": "tag",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "categoryId",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "includeDescendants",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "uncategorized",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "ownerId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "authorId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "includeDeleted",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "key"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "title"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "status"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "priority"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "createdAt"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "updatedAt"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "asc"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "desc"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "direction",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "minimum": 0,
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "csv"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "json"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "format",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/projects/{projectId}/requirements/reorder": {
      "post": {
        "summary": "Reorder siblings within a category (FR-ORG-014)",
        "tags": [
          "requirements"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "categoryId",
                  "orderedIds"
                ],
                "properties": {
                  "categoryId": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "orderedIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/requirements/{requirementId}/purge": {
      "post": {
        "summary": "Permanently purge a long-deleted requirement (FR-REQ-014)",
        "tags": [
          "requirements"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "requirementId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
...
}